- changed growStorage functions to use (unsigned long long) types

+++ _ocCore.h
@@ -9,15 +9,14 @@
 
 #define fmax(a, b) ((a) > (b) ? (a) : (b))
 
-#define growStorage(old, osize, factor) (_growStorage((old), (osize), (factor), __FILE__, __LINE__))
+#define growStorage(old, osize, factor) (_growStorage((old), (unsigned long long) (osize), (factor), __FILE__, __LINE__))
 
 /**
  * growStorage - increase the storage for a block by the given factor, copying the
  * old data into the new block. This function is called via the macro above, so
  * that memory usage can be tracked.
  */
-extern void *_growStorage(void *old, long oldSize, long factor,
-			  const char *file, long line);
+extern void *_growStorage(void *old, unsigned long long oldSize, long factor, const char *file, long line);
 
 /**
  * Determine if two variable lists are the same
